Overhaul of the header component. It has a service for some of its logic. It can now be configured (via header-settings in settings.ts) to: - Show logo (boolean) - show title (boolean) - show button to colofon (boolean) - choose between title lengths (short/regular) - choose between scale of content of header (s/m/l) - choose between gap heights after the header (s/m/l) - choose between horizontal positions (left sticky, left, center, right, right sticky) - choose between vertical positions (top/middle/bottom)
It adjusts the corners to rounded depending on the position. It ignores invalid positions (middle center) and it also contains logic to move the header regardless of position-settings when the viewport goes below 1600 px, since it will then overlap. Do note that with a long title you could still see some overlap. Either change your position (eg, go to top center), or adjust the viewport-breakpoints to solve this for your case.